Cycles.Aux

Documentation

goodGraphList :: Integral a => [[a]] -> Bool

trimList :: Integral a => [a] -> [a]

wrapAround :: [a] -> [a]

wrapAround appends the first element to the list

elemIndex' :: (Eq a, Integral b) => a -> [a] -> b -> b

elemIndex' is the same as elemIndex, except it allows for any Integral type as the index and has no Maybe

bool :: Integral a => Bool -> a

length' :: (Eq a, Integral b) => [a] -> b

length' outputs a general Integral type length' :: Integral b => [a] -> b -> b length' [] p = p length' (_:xs) p = length' xs (p+1)

replicate' :: Integral b => a -> b -> [a]

padMod :: Integral a => [a] -> a -> a -> [a]

padMod pads the end of the list with the given padding so that the final length is 0 modulo the given modulus

elemIndex'' :: (Eq a, Integral b) => a -> [a] -> b -> Maybe b

elemIndex'' is the same as elemIndex', except it allows Maybe

getElem :: Integral b => [a] -> b -> a

formatByDict :: String -> [(String, String)] -> String

formatByDict takes a string and a dictionary composed of [(key_string, result_string)] and returns the string with the appropriate key replaced (if any) ((only replaces one key))

removeExtraNewlines :: String -> String

padStrLeft0s :: Integral a => String -> a -> String

s is the string, l is how long to make the final one

digLen10 :: (Num a, Integral a, Eq a) => a -> a

Number of decimal digits in n

otherElem :: Eq a => a -> [a] -> a

Get first element of l that's not equal to a (assuming no repetitons)

unlines' :: [[Char]] -> [Char]

trim :: [a] -> [a]

hStringToCList :: String -> String

enum :: [a] -> [(Int, a)]