Cycles.Forcy

Documentation

generateMaxCyCode :: [[Int]] -> [[Int]] -> Int -> Int -> Int -> (String, String)

generateMaxCyCodeAtStart :: [[Int]] -> [[Int]] -> Int -> Int -> (Int -> String, String)

generateFindCyCode :: [[Int]] -> [Char] -> Bool -> Bool -> [Char]

Following is the generator for a one-time use C program to find all cycles in a graph.

The general program will work with directed and undirected graphs and will find all *directed* cycles in the graph I plan to modify the original program to work with graphs and digraphs and each modification to optinally only count cycles instead of listing them This cannot be solved in O(V+E) time because even planar graphs can have O(2^v) cycles and listing/counting a graph cycle takes at least 1 CPU cycle

The algorithm performs a depth first search from an arbitrary vertex, marking seen vertices on the way, and only outputting cycles through the initial vertex. As all cycles through that vertex have been found by the end, it is removed and the process repeated. This way, there is no need to check if a cycle has been found yet and there are no sources of repetition.

The only weaknesses of this algorithm that I know are: 1) It requires full out-adjacency lists for each vertex (best fit for directed, sparse graphs) 2) For undirected graphs, if the cycles are considered to be undirected, it outputs one "forward" and one "backward" copy of each cycle N.B. For the purposes of inputting the cycles into the Max-Cycles part of the code, this is a strength as the algorithm needs both 3) Although the output file is formatted, it is annoying to integrate generating, compiling, running, and parsing into a single function 4) The generated code is in C. Yeah, I said it. On the plus side, unless you're running this on very big graphs (where it would take an obscene amount of time to finish), you shouldn't run into overflow issues.

The biggest plus? This algorithm is stupid-fast. Last test I ran outputted ~1.5GB of cycles from ~2k strongly-connected graphs, |E|<=36 in under 30s.

graph (edgelist), cfilename, are assumed variables

graphToNumCycles :: [[Int]] -> Bool -> IO Int

This function takes a graph and whether it is a digraph and returns the number of directed cycles in it

graphToCycles :: [[Int]] -> Bool -> IO [[Int]]

This function takes a graph and whether it is a digraph and returns a list of all directed cycles (somewhat slow, because it uses Haskell's read function).

graphToMaxcyCode :: [[Int]] -> Int -> [Char] -> IO [Char]

This function takes a graph and automatically finds all cycles to output a string containing the C code to compute its maximally cyclic orientations

compileAllInDir :: [Char] -> IO ()

This function does exactly what it says, for '.c' files.

runAllInDir :: [Char] -> IO ()

trim :: [a] -> [a]