Cycles.Maxcy

Documentation

convIndivCycle :: (Integral a, Bits a) => [[a]] -> [a] -> [(Bool, a)]

elist is edge list and vlist is cycle vertex list

aVarList :: (Integral a, Bits a, Integral b, Bits b) => [[(Bool, a)]] -> a -> [b]

aVarList is a[var][j], likewise for bVarList

bVarList :: (Integral a, Bits a) => [[(Bool, a)]] -> a -> [a]

cVarList :: (Integral a, Bits a) => [a] -> [a] -> [a]

Note: (1-x) == ~x for x <- [0,1]

dVarList :: (Integral a, Bits a) => [a] -> [a] -> [a]

listToULL :: Integral a => [a] -> CULLong

listToULL converts a length <= 64 list of 0,1 values into a single CULLong

listToULLs :: Integral a => [a] -> [CULLong]

listToULLs converts a list of length > 64 to a list of CULLong values

addULLs :: String -> String

listFirstAnd :: [[CULLong]] -> [[CULLong]] -> [[CULLong]] -> [[CULLong]]

listFirstAnd replaces a!!0 with b!!0 .&. c!!0

sndMap2 :: [[(a, b)]] -> [[b]]

maxPositive :: (Ord a, Num a) => [a] -> a

maxPositiveSnd :: (Num a, Ord b, Num b) => [(a, b)] -> (a, b)

listsMaxPositive :: (Ord a, Num a) => [[a]] -> a

tallyElem :: Eq a => [[a]] -> a -> Int

This function counts the number of occurances of x in lists

aMostOftenElem :: [[(a, Int)]] -> Int

This function finds some most commonly occuring snd element in a list of lists

switchInTupLists :: [[(a, Int)]] -> Int -> Int -> [[(a, Int)]]

This function takes a list of lists of tuples and switches all occurances of a with b and all occurances of b with a.

frontLoadTupLists :: [[(a, Int)]] -> [[(a, Int)]]

This function makes sure that '0' is a most often snd element in a list of lists

trimCycleList :: [[(Bool, Int)]] -> [[(Bool, Int)]]

This function deletes the cycles which will be nulled out by the first edge being zero, i.e. it deletes cycles made impossible by fixing the 0th edge forward. Additionally, this function re-frontloads the lists so that more maxcy orientations "may" be found more quickly, reducing the need to print cycles. (True, _) if edge is forward in cycle, blist[][] = 0 (is how it's shown to be forward)

generateMaxCyCode :: [[Int]] -> [[Int]] -> Int -> Int -> Int -> (String, String)

generateMaxCyCodeAtStart :: [[Int]] -> [[Int]] -> Int -> Int -> (Int -> String, String)

makeAsmCounter4 :: Int -> [Char]

makeAsmCounter2 :: Int -> [Char]

makeAsmCounter1 :: Int -> [Char]

makeAsmCounterN :: Int -> [Char]

makeAsmCounter :: Int -> String

listFromULL :: CULLong -> String

printULLs :: [[CULLong]] -> String

sortBySnd :: Ord b => [(a, b)] -> [(a, b)]