Test.Testing

Documentation

maxNumCyForTest :: Int

maxNumEdgesForTest :: Int

removeDirIfExists :: FilePath -> IO ()

sort2 :: Ord a => [[a]] -> [[a]]

cutUntoChar :: [Char] -> Char -> [Char]

This function takes a string and char and returns everything up to (but not including) that char

readBins :: [Char] -> [Int]

This function converts a string like "0101010101001010" into a list of Ints

solutionStrToTup :: [Char] -> ([Int], Int)

This function converts a solution string (such as "[10101010, 2384]") to a (list of bins, number)

orientGraph :: (Ord a1, Num a1, Num a, Eq a) => [a] -> [[a1]] -> [[a1]]

This function takes a list of directions encoded as '1's and '0's ('0' is forward, i.e. [a,b] -> a<b) and a graph and returns a digraph orientGraph :: [Int] -> [[Int]] -> [[Int]] --'0' forward (a<b)

strSolutionPipe1 :: [Char] -> [[Int]] -> (([Int], Int), [[Int]])

strSolutionPipe2 :: (([Int], Int), [[Int]]) -> ([[Int]], Int)

strSolutionPipe3 :: ([[Int]], Int) -> IO Bool

strSolutionPipe :: [Char] -> [[Int]] -> IO Bool

This function takes a solution string and a graph and returns whether the solution agrees with the graphToNumCycles interpretation

checkMaxCySolution :: FilePath -> IO Bool

This function takes a file (path) and checks each solution against the graphToNumCycles interpretation

cycleGraph :: Int -> [[Int]]

completeGraph :: Int -> [[Int]]

This function returns the complete graph with n vertices

completeGraphNumCy :: Int -> Int

This function returns the number of undirected cycles in (completeGraph n) The formula is divided by k because with cycles of size k, there are k orientation of each cycle. cycles size k require three choices: n possibilities, n-1, n-2 etc. Sum[Product[(i - k + n)2 k, {i, 1, k}], {k, 3, n}] (Undirected cycles, '2' removed for directed cycles)

shuttleGraph :: Int -> [[Int]]

This function returns the shuttle graph of size n, which looks like |=|=|=| The following shows how the formula was derived: 3->1,4 4->2,3 5->3,6 6->4,5 n->n - 2, n - 1 + 2 * (mod n 2) The shuttle graph is good for debugging because of the following property: There are (n+1) cycles that are a cap or middle square (units) There are (n+0) cycles that are an adjacent pair of units There are (n-1) cycles that are an adjacent triple of units etc. Because of this property, this graph has exactly (2 + 3 n + n^2)/2 cycles (equal to TriangularNumber(n+1)). This graph has 8 + 3*n edges

shuttleGraphNumCy :: Int -> Int

Returns the number of undirected cycles in (shuttleGraph n)

wheelGraph :: Int -> [[Int]]

This function returns the wheel graph of size n Constructed by making the spokes from '0', adding all but one of the outer cycle, then adding the final edge This graph has 2*(n+1) edges

wheelGraphNumCy :: Int -> Int

For the wheel graph, starting with n == 4 (by mathematica's definition, isomorphic to K4), The nth graph is (n-1) triangles joined at a common central vertex and each joined to the next in a wheel. The number of cycles may be found as follows: ( 1 ) There are (n-1) 1-triangle cycles ( 2 ) There are (n-1) 2-triangle cycles ( . ) ... (n-2) There are (n-1) (n-2)-triangle cycles With the addition of the single cycle of all the triangles, this gives that there is a total of (n-1)*(n-2) + 1 cycles

resizeN :: Int -> Int -> Int

testNumCy :: [[Int]] -> Int -> Property

This property takes a graph and a known result for the number of undirected cycles to test graphToNumCycles

prop_goodCycleGraphNumCy :: Int -> Property

prop_goodCompleteGraphNumCy :: Int -> Property

prop_goodShuttleGraphNumCy :: Int -> Property

prop_goodWheelGraphNumCy :: Int -> Property

testMaxCy' :: [[Int]] -> Int -> IO Bool

testMaxCy :: [[Int]] -> Int -> Property

prop_goodCompleteGraphFindCy :: Int -> Int -> Property

prop_goodShuttleGraphFindCy :: Int -> Int -> Property

prop_goodWheelGraphFindCy :: Int -> Int -> Property

main :: IO ()